pyproject.toml
[tool.poetry]
name = "qrcodegenpy"
version = "0.1.0"
description = "QR Code Generator in python"
authors = ["Shan Konduru <shankonduru@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
qrcode = "^7.4.2"
pillow = "^10.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-html = "^4.1.1"
qrcode = "^7.4.2"
pillow = "^10.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
